spirv-as: Add OpUnknown pseudo-instruction (#6024)
We are adding support for vk::SpirvType to HLSL in Clang. This is a language feature that allows users to directly specify the SPIR-V opcode of a type they want to use in HLSL. Knowing the opcode, we can generate a valid binary directly. But LLVM also provides the option for producing text output, which is used for FileCheck tests that match regular expressions against that output. We do not have a way to generate SPIR-V assembly text for instructions with unknown opcodes.

We will also need this when we add support for HLSL’s vk::ext_instruction, which allows users to create functions representing arbitrary instructions.