opt: add resolve-binding-conflicts pass (#6044)
Assumes the input is a valid Vulkan module.

A conflict exists when an entry point uses two variables
with the same descriptor set and binding. Vulkan allows one kind
of conflict: where the two variables are an image (array of images)
variable and a sampler (array of samplers) variable.

The sampler variable's binding is incremented, and that increment
ripples up through higher numbered bindings until no more conflicts
remain.

Bug: crbug.com/398231475

Co-authored-by: Steven Perron <stevenperron@google.com>