[OPT] Delete decoration for OpPhi when unrolling
When unrolling, the decorations on an OpPhi in the header of the loop
are treated like any other user outside of the loop. The id of the OpPhi
is replaced by the value in the final iteration. This causes the
decorations to be applied to this other value, which is not
neccessarially correct.

Contributes to https://github.com/microsoft/DirectXShaderCompiler/issues/6914