[opt] Move debug instruction when neccessary in copy prop arrays.
Debug instructions for a function scope variable may appear before the
store. We use the store as the insertion point for a new access change
to the source variable assuming all uses of the variable are dominated
by the store. This causes problems for the debug instruction. It will
then receive an operand that has not been defined.

The fix this we move the debug instruction to after the new access chain
if necessary. We opt for this option because:

1. We did not want to avoid doing the optimization because of debug
   information.
2. There is no other reasonable place to insert the access chain. If we
   move it to the start of the entry block, some operands may not be
defined, and may not be available. If we put it anywhere else, we
potentially get the same problem as we currently have with the store.
3. Moving the debug value or debug declare instructions are not a
   problem. Before the store the varible was undefined, so the debug
information would not enable any useful information in the debugger
anyway.

Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/7025