[OPT] Use conservative default case for `GetPtr`
The `MemPass::GetPtr` assumed that variable pointers were not used. So
if the pointer instruction was not a variable or function parameter, it
assumed it was something like an OpAccessChain. This is not correct with
variable pointers. It could be a load or a phi or other things.

This modifies the code the have a default case that assumes the base
variable could not be found. It will only return a variable id if it
recornizes all of the instructions.

Fixes #6157