We want to use extension `SPV_KHR_variable_pointers` with vulkan memory (#6121)
model, but shader compiler crashes.

The typical case is passing the buffer to a function and use it. In
function, we need to load the real buffer from variable pointers, it's
okay.

But when we run pass `upgrade-memory-model`, we missed to handle the
`OpLoad` with variable pointers, which resulted in missing 1 dimension
to calculate the access element type. `OpStore` also has the same
problem.

Moreover, `OpLoad` and `OpStore` shouldn't be marked as
`NonPrivatePointer` if the target pointer is storage class Function or
Private.