dis: Add a new --style option
spirv-dis automatically applies some color to the output when printing
to the console. While this is useful, it has certain limitations:

* The result id colors are not helpful, since the result id is already
  obvious (left side of `=`)
* All argument ids have the same color, so that convey little
  information.
* Color is applied to the console during disassembly generation
  (especially on windows), which makes it harder for users of
  spirv-dis-the-library to take advantage of this feature.

In this change, a new --style option is added to spirv-dis (and
SPV_BINARY_TO_TEXT_OPTION_STYLE to the header) with the goal of being
able to more easily determine the type (as in SPIR-V type) and kind of
an id, while allowing stylization beyond just the console.

The styling scheme is quite simple. A few characters in between \xE
(style begin) and \xF (style end) indicate the color and emphasis. For
example:

* \xEfp\xF: Begin styling a float32-y pointer type
* \xEc\xF: Begin styling a constant value (of a non-numeric type)
* \xE\xF: End of style

With this option, `spvBinaryToText` sprinkles the output with these
markers which the caller can post-process as necessary. `spirv-dis` does
such post-processing by outputting a Legend and substituting these style
markers with various colors and emphases. A web application for example
might substitute the markers with `<style class="spirv-style-...">` and
`</style>`.