spirv-opt: Handle id overflow in merge return pass
The merge return pass can fail to create new basic blocks and
instructions if the module runs out of ids. This can lead to a crash.

This change modifies the pass to check if new ids can be created. If
not, the pass will be aborted for the current function.