spirv-opt: Handle id overflow in descriptor array access pass
The descriptor array access pass can fail to create new instructions
and basic blocks if the module runs out of ids. This can lead to a
crash.

This change modifies the pass to check if new ids can be in created. If
not, the pass will be aborted and return a failure.