spirv-opt: Handle id overflow in interface var SROA pass
The interface variable SROA pass can fail to create new variables and
instructions if the module runs out of ids. This can lead to a crash.

This change modifies the pass to check if new ids can be created. If
not, the pass will be aborted and return a failure.