spirv-opt: fix trimming pass with OpNop (#6314)
In another PR I removed 2 handlers from the opcode handler arrays
without resizing the array.
This causes the handler map to register 2 null function pointers for
the opcode `OpNop` because its opcode is `0`.

DXC can end-up generating `OpNop` in some cases, causing the compiler
to crash in those cases.
Added a test to catch this in the future.