spirv-opt: Fix OpUndef placement in ADCE
ADCE was modifying an existing instruction in order to get the OpUndef.
This would leave the OpUndef in an invalid position. Even through it is
allowed for an OpUndef to be in a function. We are moving away from
that.

This commit create a new instruction for the OpUndef, and places it in
the right place.

Fixes #6316