spirv-val: Allow tool to process directories (#6292)
This patch enables spirv-val to process a directory as its input.
The behaviour remains unchanged, except now filename is printed
alongside a line number in the error message, when a single file
is passed. When a directory is passed the tool will process all
files in the directory and its nested directories recursively.
Only files with the `.spv` extension are processed.

When in the "directory mode" the tool will process all files (no
early exit) and print validation errors as it goes. When any of
the files fails validation non-zero code is returned.

The selection of the mode (single file or directory) is done
transparently to the user based on the path passed to the tool.

This change is done to improve validation of SPIR-V modules in
SPIR-V MLIR (see #6276).