spirv-val: Improve Explicit Layout message (#6331)
closes https://github.com/KhronosGroup/SPIRV-Tools/issues/6323 and
https://github.com/KhronosGroup/SPIRV-Tools/issues/6071

New messages now look like

> [VUID-StandaloneSpirv-None-10684] Invalid explicit layout decorations
on type for operand '5[%_ptr_Private__struct_2]', the Private storage
class has a explicit layout from the Offset decoration.

and

> [VUID-StandaloneSpirv-None-10684] Invalid explicit layout decorations
on type for operand '6[%_ptr_Workgroup__arr_uint_uint_4]', the Workgroup
storage class has a explicit layout from the ArrayStride decoration.

----

Providing both the storage class and the decoration that makes it
explicit layout should be enough, the whole details around which version
make it valid or not-valid I think is overkill, the main thing is
showing a SPIR-V generator what to look for