opt: Remove double iteration of decorations
The WhileEachDecoration() function first used GetDecorationsFor() which
walks over the decorations and collects them in a vector.  It then
called a callback for each such instructions.  With this change, the
callback is called while iterating the decorations.

This has the side effect that if the callback returns early, the rest of
the decorations are never visited.  Consequently, HasDecoration() can
early out as soon as it encounters the desired decoration.

Change-Id: I12d0c16ff3fcef6355f3a26d1e1d98ee3f01ccfe