spirv-opt: Correct ADCE bug keeping dead DebugValue Value operands (#6368)
A logic error in ProcessDebugInfo(), a pass recently added to ADCE,
causes some shaders to fail optimization. When examining the Value
operand of a DebugValue instruction, we should only be reading the live
list, as this change corrects. Previously, we were using the .Set
method, which was incorrectly marking these instructions live. Since
other dependencies may have been killed, the SPIR-V is left in an
inconsistent state.

This PR partially corrects
https://github.com/microsoft/DirectXShaderCompiler/issues/7744. The
remaining issue is in scalar-replacement=0. Since this PR has more
wide-ranging benefits and I haven't root-caused the remaining problem,
I'm submitting it now.