Adding deduplication of NonWritable OpLoads
The current logic would only mark a load as read-only if either the
OpVariable had NonWritable or a Uniform was not a storage buffer.

This adds `IsVulkanReadOnlyStorageBuffer`, which does mostly the same
stuff as `IsVulkanStorageBuffer`, with the addition of testing for
NonWritable.

When used as part of `IsReadOnlyLoad` for assigning value numbers,
prevents it from being forced a unique value.

Fixes #6412