spirv-opt: Remove lingering debug-scopes. (#6416)
Currently removing the DebugInfo.100 family of extension doesn't clean
up after itself.

This results in instructions thinking they have a debug scope attached,
which manifests as a crash in `Module::ToBinary`.

During:
```cpp
        if (!between_label_and_phi_var ||
            context()
                ->get_feature_mgr()
                ->GetExtInstImportId_OpenCL100DebugInfo()) {
          // Emit DebugScope |scope| to |binary|.
          auto dbg_inst = ext_inst_debuginfo_.begin();
          scope.ToBinary(dbg_inst->type_id(), context()->TakeNextId(),
                         dbg_inst->GetSingleWordOperand(2), binary);
        }
```

Due to `ext_inst_debuginfo` now being empty and pointing to the OpNop
sentinal. So when it attempts to fetch `GetSingleWordOperand(2)`,
attempts to load an OOB operand.

Fixes #6344