spriv-opt: Adding a rule to merge select-comparisons
It seems that things like `(a ? 1u : 0u) == 1u` are surprisingly prevalent.

When the select results in constants and the comparison is constant, we can
merge the result directly into the select.

The following cases have been added:
```
  ((a ? C0 : C1) == C2)  =  ((a ? (C0 == C2) : (C1 == C2))
  ((a ? C0 : C1) != C2)  =  ((a ? (C0 != C2) : (C1 != C2))
  ((a ? C0 : C1) <  C2)  =  ((a ? (C0 <  C2) : (C1 <  C2))
  ((a ? C0 : C1) <= C2)  =  ((a ? (C0 <= C2) : (C1 <= C2))
  ((a ? C0 : C1) >  C2)  =  ((a ? (C0 >  C2) : (C1 >  C2))
  ((a ? C0 : C1) >= C2)  =  ((a ? (C0 >= C2) : (C1 >= C2))
  ((a ? C0 : C1) || C2)  =  ((a ? (C0 || C2) : (C1 || C2))
  ((a ? C0 : C1) && C2)  =  ((a ? (C0 && C2) : (C1 && C2))
```

This further opens the door to further collapses like: `(a ? true : false) = a`