spirv-opt: fix adce when ns debug info present (#6217) (#6420)
This PR fixes two problems in ADCE while processing NonSemantic debug
info. The first change is to treat DebugDeclare similarly to DebugValue:
if a Variable is dead, replace it with OpUndef. This mirrors the
treatment of DebugValue for the Value operand. This fixes
https://github.com/KhronosGroup/SPIRV-Tools/issues/6217.

Next, we declare all DebugLocalVariable, DebugExpression, and
DebugOperation instructions to be live at initialization since there is
no reason to ever kill them. They do not have variables that, if
declared live, would have follow-on impact to DCE. This fixes
https://github.com/microsoft/DirectXShaderCompiler/issues/7874.